+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## [0.20.2] - 2021-11-25
+## [0.22.0] - 2021-11-25
 ### Added
  - dos2unix/unix2dos command.
  - expand/unexpand command.
  - id command: Because GroupIds requires cgo, id command does not work docker environment.
+ - groups command.
  - whoami command.
  ### Changed
  - cowsay command to receive data from PIPE.